-Fix validation issue -Fix no sessions issues -Add delete all sessions option -Fix ui/ux issues -Fix performance issue that read the record each time in the while loop.

+++ CodingTracker.basemkasem/CodingTracker/SessionController.cs
@@ -53,43 +53,43 @@ internal class SessionController
 
     private static void UpdateStartTime(int recordId)
     {
-        AnsiConsole.Markup("[bold]Update start time[/]\n[yellow]" +
-            "Make sure to right time in this format (dd-MM-yyyy hh:mm) or (hh:mm)[/]\n");
-        DateTime startTime = PromptForTime("Enter start time: ");
-        while (startTime > GetRecord(recordId).EndTime)
+        CodingSession session = GetRecord(recordId);
+        AnsiConsole.Markup($"[bold]Update start time from {session.StartTime}[/]\n[yellow]" +
+            "Make sure to right time in this format (dd/MM/yyyy hh:mm)[/]\n");
+        DateTime newStartTime = PromptForTime("Enter start time: ");
+        while (newStartTime > session.StartTime)
         {
             AnsiConsole.Markup("[red]Start time cannot be later than end time. Please enter a valid end time.[/]\n");
-            startTime = PromptForTime("Enter start time (dd-MM-yyyy hh:mm): ");
+            newStartTime = PromptForTime("Enter start time (dd-MM-yyyy hh:mm): ");
         }
-
-        var updatedSession = GetRecord(recordId);
-        updatedSession.StartTime = startTime;
+        
+        session.StartTime = newStartTime;
 
         var sql = @"UPDATE CodingSessions 
                     SET StartTime = @startTime, Duration = @duration
                     WHERE Id = @id;";
 
-        DatabaseHelper.Connection.Execute(sql, new { startTime = startTime.ToString(), id = recordId, duration = updatedSession.Duration });
+        DatabaseHelper.Connection.Execute(sql, new { startTime = newStartTime.ToString(), id = recordId, duration = session.Duration });
     }
 
     private static void UpdateEndTime(int recordId)
     {
-        AnsiConsole.Markup("[bold]Update end time[/]\n[yellow]" +
-            "Make sure to right time in this format (dd-MM-yyyy hh:mm) or (hh:mm)[/]\n");
-        DateTime endTime = PromptForTime("Enter end time (hh:MM): ");
-        while (endTime < GetRecord(recordId).StartTime)
+        CodingSession session = GetRecord(recordId);
+        AnsiConsole.Markup($"[bold]Update end time from {session.EndTime}[/]\n[yellow]" +
+            "Make sure to right time in this format (dd/MM/yyyy hh:mm)[/]\n");
+        DateTime newEndTime = PromptForTime("Enter end time: ");
+        
+        while (newEndTime < session.EndTime)
         {
             AnsiConsole.Markup("[red]End time cannot be earlier than start time. Please enter a valid end time.[/]\n");
-            endTime = PromptForTime("Enter end time (hh:MM): ");
+            newEndTime = PromptForTime("Enter end time (hh:MM): ");
         }
-
-        var updatedSession = GetRecord(recordId);
-        updatedSession.EndTime = endTime;
+        session.EndTime = newEndTime;
 
         var sql = @"UPDATE CodingSessions 
                     SET EndTime = @endTime, Duration = @duration
                     WHERE Id = @id";
-        DatabaseHelper.Connection.Execute(sql, new { endTime = endTime.ToString(), id = recordId, duration = updatedSession.Duration });
+        DatabaseHelper.Connection.Execute(sql, new { endTime = newEndTime.ToString(), id = recordId, duration = session.Duration });
     }
 
     public static void ShowSessions()
@@ -99,6 +99,12 @@ internal class SessionController
     }
     public static void UpdateSession()
     {
+        var sessionsList = DatabaseHelper.TableList();
+        if (sessionsList.Count == 0)
+        {
+            AnsiConsole.Markup("[red]No records found. You need to add a session first.[/]\nPress any key to continue...");
+            return;
+        }
         var updateRecord = AnsiConsole.Prompt(new SelectionPrompt<CodingSession>()
         .Title("Which record will you update?")
         .AddChoices(DatabaseHelper.TableList())
@@ -124,15 +130,38 @@ internal class SessionController
 
     public static void DeleteSession()
     {
-        var deleteRecord = AnsiConsole.Prompt(
-            new SelectionPrompt<CodingSession>()
-            .Title("Which record will you delete?")
-            .AddChoices(DatabaseHelper.TableList())
-            );
-        var sql = @"DELETE FROM CodingSessions 
-                    WHERE Id = @id";
-        DatabaseHelper.Connection.Execute(sql, new { id = deleteRecord.Id });
-        AnsiConsole.Markup("[green]Record is deleted successfully.[/]\nPress any key to continue...");
+        var sessionsList = DatabaseHelper.TableList();
+        if (sessionsList.Count == 0)
+        {
+            AnsiConsole.Markup("[red]No records found. You need to add a session first.[/]\nPress any key to continue...");
+            return;
+        }
 
+        var updateOptionInput = AnsiConsole.Prompt(
+            new SelectionPrompt<DeleteOptions>()
+            .Title("What would you like to do?")
+        .AddChoices(Enum.GetValues<DeleteOptions>())
+        );
+
+        switch (updateOptionInput)
+        {
+            case DeleteOptions.DeleteRecord:
+                var deleteRecord = AnsiConsole.Prompt(
+                    new SelectionPrompt<CodingSession>()
+                    .Title("Which record will you delete?")
+                    .AddChoices(sessionsList)
+                    );
+                var sql = @"DELETE FROM CodingSessions 
+                            WHERE Id = @id";
+                DatabaseHelper.Connection.Execute(sql, new { id = deleteRecord.Id });
+                AnsiConsole.Markup("[green]Record is deleted successfully.[/]\nPress any key to continue...");
+                break;
+
+            case DeleteOptions.DeleteAll:
+                var sqlDeleteAll = "DELETE FROM CodingSessions;";
+                DatabaseHelper.Connection.Execute(sqlDeleteAll);
+                AnsiConsole.Markup("[green]All record are deleted successfully.[/]\nPress any key to continue...");
+                break;
+        }
     }
 }
